refactor(n0b): split ai commands into ai_*_cmd modules
Replace monolithic ai_cmd.py with ai_image_cmd, ai_video_cmd,
ai_audio_cmd, ai_speak_cmd, ai_transcribe_cmd, and ai_research_cmd.
Shared hint/replacement helpers move to ai_common.py; video_cmd.py
stays for non-AI ffmpeg utilities.
