+++ apps/n0b/commands/ai_research_cmd.py
@@ -1,0 +1,13 @@
+"""n0b ai research — OpenAI deep research."""
+from __future__ import annotations
+
+import sys
+
+from research import run_research
+
+
+def cmd_research(args: list[str]) -> int:
+    if not args:
+        print("Usage: n0b ai research <prompt...>", file=sys.stderr)
+        return 2
+    return run_research(args)
