+++ apps/n0b/commands/ai_transcribe_cmd.py
@@ -1,0 +1,147 @@
+"""n0b ai transcribe — local Whisper."""
+from __future__ import annotations
+
+import re
+import shutil
+import subprocess
+import sys
+from pathlib import Path
+
+from ai_venv import ensure_whisper
+from commands.ai_common import (
+    parse_cli_pairs,
+    read_hints,
+    read_pair_file,
+    save_hints,
+    save_pair_file,
+    split_cli_hints,
+)
+
+HINTS_FILE = Path.home() / ".config" / "n0b" / "transcribe-hints.txt"
+REPLACEMENTS_FILE = Path.home() / ".config" / "n0b" / "transcribe-replacements.txt"
+
+_WHISPER_SNIPPET = """\
+import sys
+import whisper
+
+audio, model_name, language, prompt = sys.argv[1:5]
+print(f"loading model {model_name}...", file=sys.stderr)
+model = whisper.load_model(model_name)
+print(f"transcribing (language: {language or 'auto-detect'})...", file=sys.stderr)
+result = model.transcribe(
+    audio,
+    language=language or None,
+    initial_prompt=prompt or None,
+    fp16=False,
+    verbose=False,
+)
+print(result["text"].strip())
+"""
+
+
+def read_replacements(replacements_file: Path) -> list[tuple[str, str]]:
+    return read_pair_file(replacements_file, "n0b ai transcribe")
+
+
+def parse_cli_replacements(cli_replaces: list[str]) -> list[tuple[str, str]]:
+    return parse_cli_pairs(cli_replaces, "n0b ai transcribe")
+
+
+def apply_replacements(
+    text: str, pairs: list[tuple[str, str]]
+) -> tuple[str, list[str]]:
+    applied: list[str] = []
+    for pattern, correction in pairs:
+        def annotate(m: re.Match[str]) -> str:
+            return f"{m.group(0)} (possible transcribe error, might be '{correction}')"
+
+        try:
+            text, count = re.subn(pattern, annotate, text)
+        except re.error as exc:
+            print(
+                f"n0b ai transcribe: bad replacement regex {pattern!r}: {exc}",
+                file=sys.stderr,
+            )
+            continue
+        if count:
+            applied.append(f"{pattern} => {correction} (x{count})")
+    return text, applied
+
+
+def save_replacements(cli_replaces: list[str], replacements_file: Path) -> int:
+    return save_pair_file(cli_replaces, replacements_file, "n0b ai transcribe")
+
+
+def cmd_transcribe(
+    audio: str | None,
+    hints: list[str],
+    language: str | None,
+    model: str,
+    save: bool = False,
+    replaces: list[str] | None = None,
+) -> int:
+    replaces = replaces or []
+    if save:
+        saved = False
+        if split_cli_hints(hints):
+            save_hints(hints, HINTS_FILE)
+            saved = True
+        if parse_cli_replacements(replaces):
+            save_replacements(replaces, REPLACEMENTS_FILE)
+            saved = True
+        if not saved:
+            print(
+                "n0b ai transcribe: --save needs at least one --hint or --replace",
+                file=sys.stderr,
+            )
+            return 2
+        if audio is None:
+            return 0
+    if audio is None:
+        print("n0b ai transcribe: audio file required (or --save)", file=sys.stderr)
+        return 2
+    path = Path(audio).expanduser()
+    if not path.is_file():
+        print(f"n0b ai transcribe: no such file: {audio}", file=sys.stderr)
+        return 1
+    if shutil.which("ffmpeg") is None:
+        print(
+            "n0b ai transcribe: ffmpeg not found (try: brew install ffmpeg)",
+            file=sys.stderr,
+        )
+        return 1
+    file_hints = read_hints(HINTS_FILE)
+    cli_hints = split_cli_hints(hints)
+    prompt = ", ".join(file_hints + cli_hints)
+    if prompt:
+        print(
+            f"hints: {prompt}\n"
+            f"  ({len(file_hints)} from {HINTS_FILE}, {len(cli_hints)} from --hint)",
+            file=sys.stderr,
+        )
+    else:
+        print(
+            f"hints: none (create {HINTS_FILE}, or pass --hint, add --save to keep)",
+            file=sys.stderr,
+        )
+    pairs = read_replacements(REPLACEMENTS_FILE) + parse_cli_replacements(replaces)
+    if pairs:
+        print(f"replacements: {len(pairs)} pattern(s) loaded", file=sys.stderr)
+    try:
+        python = ensure_whisper()
+    except subprocess.CalledProcessError as exc:
+        print(f"n0b ai transcribe: Whisper setup failed: {exc}", file=sys.stderr)
+        return 1
+    proc = subprocess.run(
+        [str(python), "-c", _WHISPER_SNIPPET, str(path), model, language or "", prompt],
+        stdout=subprocess.PIPE,
+        text=True,
+    )
+    if proc.returncode != 0:
+        return proc.returncode
+    text, applied = apply_replacements(proc.stdout.strip(), pairs)
+    if pairs:
+        note = "; ".join(applied) if applied else "none matched"
+        print(f"replacements applied: {note}", file=sys.stderr)
+    print(text)
+    return 0
